Decrease the listener priority
In Symfony's convention, the listeners priorities are from -255 to 255, therefore a priority of 10000 is... a bit high. The main problem of not being in the conventions is that it also disallow other listeners to be the first ones, without knowing about this bundle.


I changed it to 250 so this is an extremely high value that still allows other listeners to be placed before this one (not technically but in the priority conventions).